requirements.txt
psutil
redis
psycopg2-binary
ping3
python-dotenv
pytest
